# Run FrequencyFilter only when necessary